added user information security to user_routes

+++ app/api/user_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, jsonify
-from flask_login import login_required
+from flask_login import login_required, current_user
 from app.models import User, db
 
 user_routes = Blueprint('users', __name__)
@@ -28,6 +28,9 @@ def update_user(id):
     """
     Update a user profile by ID.
     """
+    if current_user.id != id:
+        return jsonify({'errors': 'You do not have permission to update this user.'}), 403
+
     user = User.query.get_or_404(id)
     data = request.get_json()
     user.username = data.get('username', user.username)
@@ -44,6 +47,9 @@ def delete_user(id):
     """
     Delete a user by ID.
     """
+    if current_user.id != id:
+        return jsonify({'errors': 'You do not have permission to delete this user.'}), 403
+
     user = User.query.get_or_404(id)
     db.session.delete(user)
     db.session.commit()
